Extract terminal search into composed TerminalSearch object
Move the search bar widgets, state, and the ~20 search methods (sync VTE +
async PyXterm paths, key controllers) out of TerminalWidget into a new
terminal_search.py TerminalSearch(terminal) that the widget owns as
self._search. Behavior-preserving.

TerminalWidget keeps thin forwarders for the names external code binds:
_show_search_overlay, _hide_search_overlay, handle_search_result(s),
search_text, and the search_revealer reference (window.py, the backend owner
callbacks, the search GAction). Characterization tests updated to target the
extracted object.
