+++ sshpilot/terminal.py
@@ -46,6 +46,7 @@ from .terminal_color_utils import (
     relative_luminance,
     get_contrast_color,
 )
+from .terminal_search import TerminalSearch
 
 
 # Substrings (lowercased) in terminal output that mean an ssh attempt is failing.
@@ -187,13 +188,9 @@ class TerminalWidget(Gtk.Box):
         # Create backend first before setup
         self._shortcut_controller = None
         self._scroll_controller = None
-        self._search_key_controller = None
         self._config_handler = None
         self._supported_encodings = None
         self._updating_encoding_config = False
-        self._last_search_text = ''
-        self._last_search_case_sensitive = False
-        self._last_search_regex = False
         try:
             self._pass_through_mode = bool(self.config.get_setting('terminal.pass_through_mode', False))
         except Exception:
@@ -225,103 +222,10 @@ class TerminalWidget(Gtk.Box):
         self.overlay = Gtk.Overlay()
         self.overlay.set_child(self.scrolled_window)
 
-        # Search overlay elements (revealer styled like other banners)
-        self.search_revealer = Gtk.Revealer()
-        self.search_revealer.set_reveal_child(False)
-        self.search_revealer.set_transition_type(Gtk.RevealerTransitionType.SLIDE_DOWN)
-        self.search_revealer.set_halign(Gtk.Align.FILL)
-        self.search_revealer.set_valign(Gtk.Align.START)
-        self.search_revealer.set_hexpand(True)
-
-        search_banner = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        search_banner.add_css_class('banner')
-        search_banner.set_hexpand(True)
-
-        search_header = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-        search_header.set_margin_start(12)
-        search_header.set_margin_end(12)
-        search_header.set_margin_top(8)
-        search_header.set_margin_bottom(4)
-
-        search_title = Gtk.Label(label=_("Search Terminal"))
-        search_title.set_xalign(0)
-        search_title.set_hexpand(True)
-        search_title.add_css_class('title-4')
-        search_header.append(search_title)
-
-        from sshpilot import icon_utils
-        self.search_close_button = Gtk.Button()
-        icon_utils.set_button_icon(self.search_close_button, 'window-close-symbolic')
-        self.search_close_button.add_css_class('flat')
-        self.search_close_button.set_valign(Gtk.Align.CENTER)
-        self.search_close_button.connect('clicked', lambda *_: self._hide_search_overlay())
-        self.search_close_button.set_tooltip_text(_("Close terminal search"))
-        search_header.append(self.search_close_button)
-
-        search_banner.append(search_header)
-
-        search_controls = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        search_controls.set_margin_start(12)
-        search_controls.set_margin_end(12)
-        search_controls.set_margin_bottom(8)
-
-        self.search_entry = Gtk.SearchEntry()
-        self.search_entry.set_placeholder_text(_("Search terminal history"))
-        self.search_entry.set_hexpand(True)
-        self.search_entry.connect('search-changed', self._on_search_entry_changed)
-        self.search_entry.connect('activate', self._on_search_entry_activate)
-        self.search_entry.connect('stop-search', self._on_search_entry_stop)
-
-        entry_key_controller = Gtk.EventControllerKey()
-        entry_key_controller.connect('key-pressed', self._on_search_entry_key_pressed)
-        self.search_entry.add_controller(entry_key_controller)
-
-        search_controls.append(self.search_entry)
-
-        self.search_count_label = Gtk.Label(label="")
-        self.search_count_label.add_css_class("dim-label")
-        self.search_count_label.set_width_chars(7)
-        self.search_count_label.set_xalign(1)
-        search_controls.append(self.search_count_label)
-
-        self.search_prev_button = Gtk.Button()
-        icon_utils.set_button_icon(self.search_prev_button, 'go-up-symbolic')
-        self.search_prev_button.set_tooltip_text(_("Find previous match"))
-        self.search_prev_button.connect('clicked', self._on_search_previous)
-        self.search_prev_button.set_sensitive(False)
-        search_controls.append(self.search_prev_button)
-
-        self.search_next_button = Gtk.Button()
-        icon_utils.set_button_icon(self.search_next_button, 'go-down-symbolic')
-        self.search_next_button.set_tooltip_text(_("Find next match"))
-        self.search_next_button.connect('clicked', self._on_search_next)
-        self.search_next_button.set_sensitive(False)
-        search_controls.append(self.search_next_button)
-
-        search_banner.append(search_controls)
-        self.search_revealer.set_child(search_banner)
-
-        # Install CSS for search banner to ensure solid background
-        try:
-            display = Gdk.Display.get_default()
-            if display and not getattr(display, '_sshpilot_search_banner_css_installed', False):
-                css_provider = Gtk.CssProvider()
-                css_provider.load_from_data(b"""
-                    .search-banner {
-                        background-color: @headerbar_bg_color;
-                        color: @headerbar_fg_color;
-                        border-bottom: 1px solid @borders;
-                    }
-                """)
-                Gtk.StyleContext.add_provider_for_display(
-                    display, css_provider, Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION
-                )
-                setattr(display, '_sshpilot_search_banner_css_installed', True)
-        except Exception:
-            pass
-
-        # Add the search-banner CSS class to ensure solid background
-        search_banner.add_css_class('search-banner')
+        # Search overlay (widgets + state + backend driving) live in a composed
+        # object; TerminalWidget keeps thin forwarders for the external API.
+        self._search = TerminalSearch(self)
+        self.search_revealer = self._search.search_revealer
 
         # Connecting overlay elements
         self.connecting_bg = Gtk.Box()
@@ -3444,7 +3348,7 @@ class TerminalWidget(Gtk.Box):
             logger.debug(f"Failed to install shortcuts: {e}")
 
         try:
-            self._ensure_search_key_controller()
+            self._search._ensure_search_key_controller()
         except Exception:
             pass
     
@@ -3494,51 +3398,6 @@ class TerminalWidget(Gtk.Box):
         except Exception as e:
             logger.debug(f"Failed to setup mouse wheel zoom: {e}")
 
-    def _ensure_search_key_controller(self):
-        """Attach the search shortcut controller to the terminal if needed."""
-        if getattr(self, '_search_key_controller', None) is not None:
-            return
-
-        try:
-            controller = Gtk.EventControllerKey()
-            controller.connect('key-pressed', self._on_vte_search_key_pressed)
-            if self.vte is not None:
-                self.vte.add_controller(controller)
-            elif self.terminal_widget is not None:
-                self.terminal_widget.add_controller(controller)
-            self._search_key_controller = controller
-            logger.debug("Search key controller installed")
-        except Exception as exc:
-            logger.debug("Failed to install search key controller: %s", exc)
-
-    def _on_vte_search_key_pressed(self, controller, keyval, keycode, state):
-        """Handle global terminal search shortcuts on the VTE widget."""
-        try:
-            shift = bool(state & Gdk.ModifierType.SHIFT_MASK)
-            primary = bool(state & Gdk.ModifierType.CONTROL_MASK)
-            meta = bool(state & Gdk.ModifierType.META_MASK)
-
-            if keyval in (Gdk.KEY_f, Gdk.KEY_F) and (primary or meta):
-                if hasattr(self, 'search_revealer') and self.search_revealer.get_reveal_child():
-                    self._hide_search_overlay()
-                else:
-                    self._show_search_overlay(select_all=True)
-                return True
-
-            if keyval in (Gdk.KEY_g, Gdk.KEY_G) and (primary or meta):
-                if shift:
-                    self._on_search_previous()
-                else:
-                    self._on_search_next()
-                return True
-
-            if keyval == Gdk.KEY_Escape and hasattr(self, 'search_revealer') and self.search_revealer.get_reveal_child():
-                self._hide_search_overlay()
-                return True
-        except Exception as exc:
-            logger.debug("Terminal search key handling failed: %s", exc)
-        return False
-
     def _remove_custom_shortcut_controllers(self):
         """Detach any custom shortcut or scroll controllers from the VTE widget."""
         ctrl = getattr(self, '_shortcut_controller', None)
@@ -3561,15 +3420,8 @@ class TerminalWidget(Gtk.Box):
             finally:
                 self._scroll_controller = None
 
-        search_ctrl = getattr(self, '_search_key_controller', None)
-        if search_ctrl is not None:
-            try:
-                if hasattr(self.vte, 'remove_controller'):
-                    self.vte.remove_controller(search_ctrl)
-            except Exception as exc:
-                logger.debug("Failed to remove search key controller: %s", exc)
-            finally:
-                self._search_key_controller = None
+        if getattr(self, '_search', None) is not None:
+            self._search.teardown_key_controller()
 
     def _apply_pass_through_mode(self, enabled: bool):
         """Enable or disable custom shortcut handling based on configuration."""
@@ -4299,316 +4151,24 @@ class TerminalWidget(Gtk.Box):
         except Exception as e:
             logger.error(f"Failed to reset terminal zoom: {e}")
 
-    def _apply_search_highlight_colors(self):
-        """Switch VTE highlight to a high-contrast color while search is active."""
-        try:
-            if self.vte is None:
-                return
-            search_bg = Gdk.RGBA()
-            search_bg.parse('#F4D03F')  # Bright amber — visible on dark and light backgrounds
-            search_fg = Gdk.RGBA()
-            search_fg.parse('#000000')
-            if hasattr(self.vte, 'set_color_highlight'):
-                self.vte.set_color_highlight(search_bg)
-            if hasattr(self.vte, 'set_color_highlight_foreground'):
-                self.vte.set_color_highlight_foreground(search_fg)
-        except Exception as exc:
-            logger.debug("Failed to apply search highlight colors: %s", exc)
+    # --- Search: forwarders to the composed TerminalSearch (self._search) ---
+    # External code binds these names (window.py, the backend `owner`
+    # callbacks, the "search" GAction), so keep them resolving on the widget.
 
     def _show_search_overlay(self, select_all: bool = False):
-        """Reveal the terminal search overlay and focus the search entry."""
-        try:
-            if not hasattr(self, 'search_revealer') or not self.search_revealer:
-                return
-            self._apply_search_highlight_colors()
-            self.search_revealer.set_reveal_child(True)
-            if hasattr(self, 'search_entry') and self.search_entry:
-                if select_all:
-                    try:
-                        self.search_entry.select_region(0, -1)
-                    except Exception:
-                        pass
-                self.search_entry.grab_focus()
-                self._set_search_navigation_sensitive(bool(self.search_entry.get_text()))
-        except Exception as exc:
-            logger.debug("Failed to show search overlay: %s", exc)
+        return self._search._show_search_overlay(select_all)
 
     def _hide_search_overlay(self):
-        """Hide the search overlay and return focus to the terminal."""
-        try:
-            if hasattr(self, 'search_revealer') and self.search_revealer:
-                self.search_revealer.set_reveal_child(False)
-            self._set_search_error_state(False)
-            self._update_search_count_label(-1, 0)
-            if self.backend and hasattr(self.backend, "clear_search_decorations"):
-                try:
-                    self.backend.clear_search_decorations()
-                except Exception:
-                    pass
-            if self.backend:
-                self.backend.grab_focus()
-            # Restore theme selection color now that search is closed
-            self._apply_cursor_and_selection_colors()
-        except Exception as exc:
-            logger.debug("Failed to hide search overlay: %s", exc)
+        return self._search._hide_search_overlay()
 
-    def _set_search_navigation_sensitive(self, active: bool):
-        """Enable or disable navigation buttons based on active search state."""
-        try:
-            for button in (getattr(self, 'search_prev_button', None), getattr(self, 'search_next_button', None)):
-                if button is not None:
-                    button.set_sensitive(bool(active))
-        except Exception as exc:
-            logger.debug("Failed to update search navigation sensitivity: %s", exc)
+    def handle_search_result(self, *args, **kwargs):
+        return self._search.handle_search_result(*args, **kwargs)
 
-    def _set_search_error_state(self, has_error: bool):
-        """Toggle error styling on the search entry when matches are not found."""
-        entry = getattr(self, 'search_entry', None)
-        if not entry:
-            return
-        try:
-            if has_error:
-                entry.add_css_class('error')
-            else:
-                entry.remove_css_class('error')
-        except Exception:
-            pass
-
-    def _clear_search_pattern(self):
-        """Clear any active search pattern from the terminal."""
-        self._last_search_text = ''
-        self._last_search_case_sensitive = False
-        self._last_search_regex = False
-        self._search_has_match = False
-        self._set_search_navigation_sensitive(False)
-        self._set_search_error_state(False)
-        self._update_search_count_label(-1, 0)
-        try:
-            if self.backend:
-                if hasattr(self.backend, "search_set_query"):
-                    self.backend.search_set_query(None)
-                else:
-                    self.backend.search_set_regex(None)
-        except Exception:
-            pass
-
-    def _update_search_count_label(self, result_index: int, result_count: int) -> None:
-        label = getattr(self, "search_count_label", None)
-        if label is None:
-            return
-        try:
-            if result_count <= 0:
-                label.set_text("")
-            elif result_index < 0:
-                # SearchAddon: resultIndex == -1 when highlight threshold exceeded.
-                label.set_text(f"{result_count}+")
-            else:
-                label.set_text(f"{result_index + 1}/{result_count}")
-        except Exception:
-            pass
-
-    def handle_search_result(
-        self,
-        found: bool,
-        *,
-        result_index: int = -1,
-        result_count: int = 0,
-    ) -> None:
-        """PyXterm async findNext/findPrevious result (script-message)."""
-        text = ""
-        if getattr(self, "search_entry", None):
-            text = self.search_entry.get_text() or ""
-        if result_count > 0:
-            self._search_has_match = True
-        else:
-            self._search_has_match = bool(found)
-        if text:
-            self._set_search_error_state(not self._search_has_match)
-        else:
-            self._set_search_error_state(False)
-        if result_count > 0 or result_index >= 0:
-            self._update_search_count_label(result_index, result_count)
-
-    def handle_search_results(self, result_index: int, result_count: int) -> None:
-        """PyXterm SearchAddon.onDidChangeResults (decorations enabled)."""
-        text = ""
-        if getattr(self, "search_entry", None):
-            text = self.search_entry.get_text() or ""
-        self._search_has_match = result_count > 0
-        if text:
-            self._set_search_error_state(result_count <= 0)
-        self._update_search_count_label(result_index, result_count)
-
-    def _is_pyxterm_backend(self) -> bool:
-        backend = self.backend
-        if backend is None:
-            return False
-        return type(backend).__name__ in ("PyXtermTerminalBackend", "PyXtermBridgeBackend")
-
-    def _update_search_pattern(self, text: str, *, case_sensitive: bool = False, regex: bool = False,
-                                move_forward: bool = True, update_entry: bool = False) -> bool:
-        """Apply or update the search pattern on the active terminal backend."""
-        if not text:
-            self._clear_search_pattern()
-            return False
-
-        pattern_changed = (
-            text != self._last_search_text or
-            case_sensitive != self._last_search_case_sensitive or
-            regex != self._last_search_regex
-        )
-
-        try:
-            if pattern_changed:
-                self._search_has_match = False
-                self._update_search_count_label(-1, 0)
-                if self.backend:
-                    if hasattr(self.backend, "search_set_query"):
-                        # Pass the raw user term; backends apply escape/flags themselves.
-                        self.backend.search_set_query(
-                            text, case_sensitive=case_sensitive, regex=regex
-                        )
-                    elif hasattr(self.backend, "vte") and self.backend.vte:
-                        pattern = text if regex else re.escape(text)
-                        if not case_sensitive and not pattern.startswith("(?i)"):
-                            pattern = "(?i)" + pattern
-                        search_regex = Vte.Regex.new_for_search(pattern, -1, 0)
-                        self.backend.search_set_regex(search_regex)
-                        if hasattr(self.backend.vte, "search_set_wrap_around"):
-                            self.backend.vte.search_set_wrap_around(True)
-
-                self._last_search_text = text
-                self._last_search_case_sensitive = case_sensitive
-                self._last_search_regex = regex
-
-            self._set_search_navigation_sensitive(True)
-
-            if move_forward:
-                return self._run_search(True, update_entry=update_entry, from_text_change=True)
-
-            if update_entry:
-                self._set_search_error_state(False)
-
-            return True
-        except Exception as exc:
-            logger.error(f"Search failed: {exc}")
-            if update_entry:
-                self._set_search_error_state(True)
-            return False
-
-    def _run_search(self, forward: bool = True, *, update_entry: bool = False,
-                    from_text_change: bool = False) -> bool:
-        """Execute search navigation in the requested direction."""
-        try:
-            if self.backend:
-                found = self.backend.search_find_next() if forward else self.backend.search_find_previous()
-            else:
-                found = False
-        except Exception as exc:
-            logger.error(f"Search navigation failed: {exc}")
-            found = False
-
-        # PyXterm reports found/not-found asynchronously via handle_search_result.
-        if self._is_pyxterm_backend():
-            return True
-
-        if found:
-            self._search_has_match = True
-
-        if update_entry:
-            # Only show the error state when the pattern has no matches at all.
-            # When navigating (Enter / Ctrl+G), VTE may return False for a single
-            # already-highlighted match even with wrap-around enabled — that is not
-            # a "no results" condition, so don't turn the entry red.
-            no_match = not found and (from_text_change or not self._search_has_match)
-            self._set_search_error_state(no_match)
-
-        return bool(found)
-
-    def _on_search_entry_changed(self, entry):
-        """React to text edits in the search entry."""
-        text = entry.get_text() if entry else ''
-        if not text:
-            self._clear_search_pattern()
-            return
-        self._update_search_pattern(text, move_forward=True, update_entry=True)
-
-    def _on_search_entry_activate(self, entry):
-        """Handle Enter key in the search entry."""
-        self._on_search_next()
-
-    def _on_search_entry_stop(self, entry):
-        """Handle stop-search events (Escape or clear button)."""
-        if not entry.get_text():
-            self._clear_search_pattern()
-        self._hide_search_overlay()
-
-    def _on_search_entry_key_pressed(self, controller, keyval, keycode, state):
-        """Handle additional shortcuts while the search entry is focused."""
-        try:
-            shift = bool(state & Gdk.ModifierType.SHIFT_MASK)
-            primary = bool(state & Gdk.ModifierType.CONTROL_MASK)
-            meta = bool(state & Gdk.ModifierType.META_MASK)
-
-            if keyval in (Gdk.KEY_g, Gdk.KEY_G) and (primary or meta):
-                if shift:
-                    self._on_search_previous()
-                else:
-                    self._on_search_next()
-                return True
-
-            if keyval in (Gdk.KEY_f, Gdk.KEY_F) and (primary or meta):
-                self._hide_search_overlay()
-                return True
-
-            if keyval in (Gdk.KEY_Return, Gdk.KEY_KP_Enter) and shift:
-                self._on_search_previous()
-                return True
-
-            if keyval == Gdk.KEY_Escape:
-                self._hide_search_overlay()
-                return True
-        except Exception as exc:
-            logger.debug("Search entry key handling failed: %s", exc)
-        return False
-
-    def _on_search_next(self, *_args):
-        """Navigate to the next search match."""
-        text = ''
-        if hasattr(self, 'search_entry') and self.search_entry:
-            text = self.search_entry.get_text()
-
-        if text:
-            if not self._update_search_pattern(text, move_forward=False, update_entry=True):
-                return False
-        elif not self._last_search_text:
-            return False
-
-        return self._run_search(True, update_entry=True)
-
-    def _on_search_previous(self, *_args):
-        """Navigate to the previous search match."""
-        text = ''
-        if hasattr(self, 'search_entry') and self.search_entry:
-            text = self.search_entry.get_text()
-
-        if text:
-            if not self._update_search_pattern(text, move_forward=False, update_entry=True):
-                return False
-        elif not self._last_search_text:
-            return False
-
-        return self._run_search(False, update_entry=True)
+    def handle_search_results(self, *args, **kwargs):
+        return self._search.handle_search_results(*args, **kwargs)
 
     def search_text(self, text, case_sensitive=False, regex=False):
-        """Search for text in terminal"""
-        return self._update_search_pattern(
-            text,
-            case_sensitive=case_sensitive,
-            regex=regex,
-            move_forward=True,
-            update_entry=False,
-        )
+        return self._search.search_text(text, case_sensitive=case_sensitive, regex=regex)
 
     def get_connection_info(self):
         """Get connection information"""
